# [MRG] Fix KMeans convergence when tol==0