--- a/sklearn/cluster/_kmeans.py
+++ b/sklearn/cluster/_kmeans.py
@@ -358,6 +358,7 @@ def _kmeans_single_elkan(X, sample_weight, centers_init, max_iter=300,
     centers_new = np.zeros_like(centers)
     weight_in_clusters = np.zeros(n_clusters, dtype=X.dtype)
     labels = np.full(n_samples, -1, dtype=np.int32)
+    labels_old = labels.copy()
     center_half_distances = euclidean_distances(centers) / 2
     distance_next_center = np.partition(np.asarray(center_half_distances),
                                         kth=1, axis=0)[1]
@@ -376,7 +377,6 @@ def _kmeans_single_elkan(X, sample_weight, centers_init, max_iter=300,
 
     init_bounds(X, centers, center_half_distances,
                 labels, upper_bounds, lower_bounds)
-    labels_old = labels.copy()
 
     for i in range(max_iter):
         elkan_iter(X, sample_weight, centers, centers_new,
@@ -507,7 +507,6 @@ def _kmeans_single_lloyd(X, sample_weight, centers_init, max_iter=300,
                 print(f"Iteration {i}, inertia {inertia}.")
 
             centers, centers_new = centers_new, centers
-            labels_old, labels = labels, labels_old
 
             if tol == 0:
                 # When tol = 0 we check that labels did not change because
@@ -515,6 +514,7 @@ def _kmeans_single_lloyd(X, sample_weight, centers_init, max_iter=300,
                 if np.array_equal(labels, labels_old):
                     strict_convergence = True
                     break
+                labels_old = labels.copy()
             else:
                 center_shift_tot = (center_shift**2).sum()
                 if center_shift_tot <= tol: