@jeremiedbb 2020-07-22
I changed to always check convergence based on labels first and then on tol. I tested the impact on performance in the worst case, i.e. many samples, few centers, few features. It makes the computation of labels comparison as costly as possible compared to the other computations. In that case the drop of performance is rather small, see below. I think it's acceptable (recalling that it's in the worst situation. I many situations the drop is indetectable).

```py
from sklearn.cluster import KMeans
from sklearn.datasets import make_blobs
import numpy as np

X = np.random.RandomState(0).random_sample((1000000, 3))

km = KMeans(n_clusters=3, tol=1e-16, n_init=1, init='random', random_state=0)

# master
%timeit km.fit(X)
2.19 s ± 11.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

km.n_iter_
71

# this pr
%timeit km.fit(X)
2.26 s ± 25.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

km.n_iter_
71
```