@ogrisel 2020-07-22
Another suggestion: the coverage report highlights the fact that the new convergence messages are not covered.

It might be interesting to extend the tests using the capsys fixture to make some assertion on those, in particular with `tol=0` or `1e-100`:

https://docs.pytest.org/en/stable/capture.html#accessing-captured-output-from-a-test-function