@jeremiedbb 2020-07-22
I added a test checking all expected outputs of KMeans in verbose mode.
I'd rather not do the same for minibatch in this pr, but rather in #17622 since #17622 changes some prints.