@ogrisel 2020-07-29
> Should this be in 0.23.2??

Seems easy to backport. Let's do it.